refactor(cli): split config module into file
